Bug 1807302 - Move onboarding related UI tests to new test class
