+++ app/src/androidTest/java/org/mozilla/fenix/ui/HomeScreenTest.kt
@@ -159,81 +159,6 @@ class HomeScreenTest {
     }
 
     @Test
-    fun dismissOnboardingUsingSettingsTest() {
-        homeScreen {
-            verifyWelcomeHeader()
-        }.openThreeDotMenu {
-        }.openSettings {
-            verifyGeneralHeading()
-        }.goBack {
-            verifyExistingTopSitesList()
-        }
-    }
-
-    @Test
-    fun dismissOnboardingUsingBookmarksTest() {
-        homeScreen {
-            verifyWelcomeHeader()
-        }.openThreeDotMenu {
-        }.openBookmarks {
-            verifyBookmarksMenuView()
-            navigateUp()
-        }
-        homeScreen {
-            verifyExistingTopSitesList()
-        }
-    }
-
-    @Test
-    fun dismissOnboardingUsingHelpTest() {
-        activityTestRule.activityRule.applySettingsExceptions {
-            it.isJumpBackInCFREnabled = false
-            it.isWallpaperOnboardingEnabled = false
-        }
-
-        homeScreen {
-            verifyWelcomeHeader()
-        }.openThreeDotMenu {
-        }.openHelp {
-            verifyHelpUrl()
-        }.goBack {
-            verifyExistingTopSitesList()
-        }
-    }
-
-    @Test
-    fun dismissOnboardingWithPageLoadTest() {
-        activityTestRule.activityRule.applySettingsExceptions {
-            it.isJumpBackInCFREnabled = false
-            it.isWallpaperOnboardingEnabled = false
-        }
-
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        homeScreen {
-            verifyWelcomeHeader()
-        }
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.goToHomescreen {
-            verifyExistingTopSitesList()
-        }
-    }
-
-    @Test
-    fun toolbarTapDoesntDismissOnboardingTest() {
-        homeScreen {
-            verifyWelcomeHeader()
-        }.openSearch {
-            verifyScanButton()
-            verifySearchEngineButton()
-            verifyKeyboardVisibility()
-        }.dismissSearchBar {
-            verifyWelcomeHeader()
-        }
-    }
-
-    @Test
     fun verifyPocketHomepageStoriesTest() {
         activityTestRule.activityRule.applySettingsExceptions {
             it.isRecentTabsFeatureEnabled = false
